Settings: reorganization
Reorganize settings categories to be more user friendly. In addition,
there is now an "Advanced mode". This mode is on by default for
userdebug and eng builds, and can be overlayed to be turned on by
default for user builds. Currently this toggle lives in About phone for
easy discoverability.

Change-Id: I18fa8c3e3ee7efc9fcaa7384d5b1cfb2495843f1
Signed-off-by: Roman Birg <roman@cyngn.com>

Conflicts:
	AndroidManifest.xml
	res/values/ose_strings.xml
	res/xml/button_settings.xml
	res/xml/dashboard_categories.xml
	res/xml/display_settings.xml
	res/xml/privacy_settings_ose.xml
	res/xml/security_settings_chooser.xml
	res/xml/security_settings_password.xml
	res/xml/security_settings_pattern.xml
	res/xml/security_settings_pin.xml
	src/com/android/settings/ButtonSettings.java
	src/com/android/settings/DeviceInfoSettings.java
	src/com/android/settings/DisplaySettings.java
	src/com/android/settings/SettingsActivity.java
	src/com/android/settings/SettingsPreferenceFragment.java